+++ src/main/java/com/example/com/netplus/config/JwtFilter.java
@@ -1,0 +1,4 @@
+package com.example.com.netplus.config;
+
+public class JwtFilter {
+}
