--- src/main/java/com/example/com/netplus/filter/JwtFilter.java
@@ -1,4 +1,0 @@
-package com.example.com.netplus.filter;
-
-public class JwtFilter {
-}
