+++ src/main/java/com/example/com/netplus/service/UserService.java
@@ -1,6 +1,6 @@
-package com.example.com.netplus.service.user;
+package com.example.com.netplus.service;
 
-import com.example.com.netplus.repository.user.UserRepository;
+import com.example.com.netplus.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
