--- src/main/java/com/example/com/netplus/utill/JwtUtill.java
@@ -1,4 +1,0 @@
-package com.example.com.netplus.utill;
-
-public class JwtUtill {
-}
